requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cloudinary==1.29.0
cryptography==37.0.4
defusedxml==0.7.1
Django==4.1
django-allauth==0.51.0
django-cloudinary-storage==0.3.0
django-gravatar2==1.4.4
django-jalali-date==1.0.1
django-rosetta==0.9.8
gunicorn==20.1.0
idna==3.3
jdatetime==4.1.0
oauthlib==3.2.0
Pillow==9.2.0
polib==1.1.1
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-decouple==3.6
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.11
whitenoise==6.2.0
